Use `__init__` rather than `handler.py`
Now the default namespace (__init__) is searched
for subclasses of Corgi, rather than the unusual
`handler` module. As long as classes are added
to __init__:
```
  file: corgi_loves_XXX/__init__.py
  from handler import SomeClass
```
then they will be picked up by the server.
